refactor: reorganize per Django Styleguide — selectors for reads, services for mutations
