+++ apps/betterangels-backend/notes/utils/__init__.py
@@ -1,3 +1,3 @@
-from accounts.utils import resolve_permission_group
+from accounts.selectors import resolve_permission_group
 
 from .note_reverter_util import NoteReverter
